requirements.txt
Django>=4.2.0
gunicorn>=21.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
anthropic>=0.7.0
requests>=2.31.0